Rewrite contact and sponsor includes
- _includes/contact.html
I've rewritten the header text for the contact section. So instead
of the text being straight from the template (which talked about
"give us a call" even though we don't have a phone number), it's
now actually relevant.

- _includes/sponsors.html
I've added a line at the bottom to write us an email if visitors
would like to sponsor us.